Adds a Hobbies section
Users will now be able to add their interests and hobbies , and have this displayed in order for other users to be able to see their unique hobbies and be able to connect with others.

This will display on their profile page, and this option is available now for new registering users.